requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
deprecation==2.1.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-RESTful==0.3.10
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.3
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
postgrest==0.15.0
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
realtime==1.0.2
six==1.16.0
sniffio==1.3.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.5
supafunc==0.3.3
typing_extensions==4.9.0
websockets==11.0.3
Werkzeug==3.0.1
zipp==3.17.0
